app/core/security.py
from datetime import datetime, timedelta, timezone
from jose import JWTError, jwt
from passlib.context import CryptContext

from app.core.config import settings

# Настройка хеширования паролей
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def verify_password(
        plain_password: str,
        hashed_password: str
) -> bool:
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    print(f"🔍 DEBUG: password type = {type(password)}")
    print(f"🔍 DEBUG: password value = {repr(password)}")
    print(f"🔍 DEBUG: password length = {len(password)}")
    if not isinstance(password, str):
        raise TypeError(f"Password must be string, got {type(password)}")
    password_bytes = password.encode('utf-8')
    print(f"🔍 DEBUG: password_bytes length = {len(password_bytes)}")

    if len(password_bytes) > 72:
        raise ValueError(f"Password is too long: {len(password_bytes)} bytes (max 72)")

    return pwd_context.hash(password)



def create_access_token(
        data: dict,
        expires_delta: timedelta | None = None
)-> str:
    to_encode = data.copy()

    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, settings.JWT_SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt


def decode_access_token(token: str) -> dict | None:
    try:
        payload = jwt.decode(token, settings.JWT_SECRET_KEY, algorithms=[settings.ALGORITHM])
        return payload
    except JWTError:
        return None


