pyproject.toml
[tool.poetry]
name = "tracker"
version = "1.0.0"
description = ""
authors = ["Daniel Zhdanov <danielzhdanov5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.46"
asyncpg = "^0.31.0"
alembic = "^1.18.3"
pydantic-settings = "^2.12.0"
psycopg2-binary = "^2.9.11"
greenlet = "^3.3.1"
python-dotenv = "^1.2.1"
python-jose = "^3.5.0"
passlib = "^1.7.4"
python-multipart = "^0.0.22"
pydantic = {extras = ["email"], version = "^2.12.5"}
bcrypt = "<4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
